# auth-service/openapi.yaml
openapi: 3.0.3
info:
  title: Supabase Auth Service API
  description: |
    A production-ready Go microservice that wraps Supabase Auth API with additional security, middleware, and comprehensive testing.

    ## Authentication

    Most endpoints require authentication via the `Authorization` header:
    ```
    Authorization: Bearer <jwt-token>
    ```

    Admin endpoints require either:
    - Service role key: `Authorization: Bearer <service-role-key>`
    - Admin user JWT: `Authorization: Bearer <admin-jwt-token>`

    ## Rate Limiting

    All endpoints include rate limiting headers:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Time until window resets (seconds)

    ## Error Handling

    All errors follow a consistent format:
    ```json
    {
      "code": 400,
      "msg": "Error description",
      "details": "Additional context (optional)"
    }
    ```

  version: 1.0.0
  contact:
    name: Supabase Auth Service
    url: https://github.com/your-org/supabase-auth-service
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://auth-api.yourdomain.com
    description: Production server
security:
  - bearerAuth: []
  - serviceRoleAuth: []
paths:
...
  /signup:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new user with email/phone and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "200":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /verify:
    get:
      tags:
        - Authentication
      summary: Verify user (redirect)
      description: Verify email/phone confirmation via GET redirect (used in email links)
      security: []
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [signup, recovery, magiclink, invite, email_change]
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: redirect_to
          in: query
          schema:
            type: string
            format: uri
      responses:
        "303":
          description: Redirect to application with tokens in fragment
          headers:
            Location:
              schema:
                type: string
                example: "https://yourapp.com#access_token=xxx&refresh_token=yyy"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags:
        - Authentication
      summary: Verify user (JSON)
      description: Verify email/phone confirmation via POST with JSON response
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyRequest"
      responses:
        "200":
          description: Verification successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /resend:
    post:
      tags:
        - Authentication
      summary: Resend confirmation
      description: Resend confirmation email or SMS
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendRequest"
      responses:
        "200":
          description: Confirmation sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message_id:
                    type: string
  /recover:
    post:
      tags:
        - Authentication
      summary: Password recovery
      description: Send password recovery email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoverRequest"
      responses:
        "200":
          description: Recovery email sent (always returns 200 to prevent email enumeration)
          content:
            application/json:
              schema:
                type: object
  /magiclink:
    post:
      tags:
        - Authentication
      summary: Send magic link
      description: Send passwordless login link via email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicLinkRequest"
      responses:
        "200":
          description: Magic link sent successfully
          content:
            application/json:
              schema:
                type: object
  /otp:
    post:
      tags:
        - Authentication
      summary: Send OTP
      description: Send one-time password via email or SMS
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OTPRequest"
      responses:
        "200":
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login
    serviceRoleAuth:
      type: http
      scheme: bearer
      description: Service role key for admin operations
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
        - msg
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 400
        msg:
          type: string
          description: Error message
          example: "Invalid request"
        details:
          type: string
          description: Additional error details
          example: "Email is required"
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        aud:
          type: string
          description: Audience
          example: "authenticated"
        role:
          type: string
          description: User role
          example: "authenticated"
        email:
          type: string
          format: email
          description: User email
          example: "user@example.com"
        phone:
          type: string
          description: User phone number
          example: "+1234567890"
        email_confirmed_at:
          type: string
          format: date-time
          description: Email confirmation timestamp
          example: "2023-01-01T00:00:00Z"
        phone_confirmed_at:
          type: string
          format: date-time
          description: Phone confirmation timestamp
          example: "2023-01-01T00:00:00Z"
        last_sign_in_at:
          type: string
          format: date-time
          description: Last sign in timestamp
          example: "2023-01-01T00:00:00Z"
        app_metadata:
          type: object
          description: Application metadata
        user_metadata:
          type: object
          description: User metadata
        created_at:
          type: string
          format: date-time
          description: User creation timestamp
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: User last update timestamp
          example: "2023-01-01T00:00:00Z"
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          description: Token type
          example: "bearer"
        expires_in:
          type: integer
          description: Token expiration in seconds
          example: 3600
        expires_at:
          type: integer
          description: Token expiration timestamp
          example: 1640995200
        refresh_token:
          type: string
          description: Refresh token
          example: "refresh_token_string"
        user:
          $ref: "#/components/schemas/User"
...
    SignupRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email
          example: "user@example.com"
        phone:
          type: string
          description: User phone number
          example: "+1234567890"
        password:
          type: string
          description: User password
          example: "password123"
        data:
          type: object
          description: Additional user metadata
        captcha_token:
          type: string
          description: Captcha token for verification
...
    VerifyRequest:
      type: object
      required:
        - type
        - token
      properties:
        type:
          type: string
          enum: [signup, recovery, magiclink, invite, email_change]
          description: Verification type
        token:
          type: string
          description: Verification token
        redirect_to:
          type: string
          format: uri
          description: Redirect URL after verification
    ResendRequest:
      type: object
      required:
        - type
        - email
      properties:
        type:
          type: string
          enum: [signup, recovery]
          description: Resend type
        email:
          type: string
          format: email
          description: User email
          example: "user@example.com"
        phone:
          type: string
          description: User phone number
          example: "+1234567890"
        captcha_token:
          type: string
          description: Captcha token for verification
    RecoverRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: User email
          example: "user@example.com"
        captcha_token:
          type: string
          description: Captcha token for verification
    MagicLinkRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: User email
          example: "user@example.com"
        create_user:
          type: boolean
          description: Whether to create user if not exists
          example: true
        captcha_token:
          type: string
          description: Captcha token for verification
    OTPRequest:
      type: object
      required:
        - phone
      properties:
        phone:
          type: string
          description: User phone number
          example: "+1234567890"
        create_user:
          type: boolean
          description: Whether to create user if not exists
          example: true
        captcha_token:
          type: string
          description: Captcha token for verification
...